DEV-9 move packet_parser serde to domain
