+++ src/Shared/PacketAdapter/domain/serde/decoder.go
