+++ src/Shared/PacketAdapter/domain/serde/encoder.go
